% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{add_sections}
\alias{add_sections}
\title{Add section lines and/or shading}
\usage{
add_sections(p, scale, shade = FALSE, ymin = NULL, ymax = NULL,
  opacity = 0.3)
}
\arguments{
\item{p}{Plot to add shading to}

\item{scale}{The number of data points in each third of the profile.
See \code{\link{get_profile}} for a description of profile and scale.}

\item{shade}{Add shading of sections or else add lines separating sections}

\item{ymin}{Minimum value of the y axis}

\item{ymax}{Maximum value of the y axis}

\item{opacity}{The opacity of the shading}
}
\value{
Ploltly object that has section shading or lines
}
\description{
add shading or lines to show the sections of a profile for a plot
}
\examples{
   profile <- get_profile(df, 300, 200, 500)
   p <- plot_profile(profile, "the best profile", sections = NULL)
   add_sections(p, scale = 500)
}
\seealso{
\code{\link{get_profile}} for a description of profile and scale.

Other plotting functions: \code{\link{get_color}},
  \code{\link{get_energy_profiles}},
  \code{\link{min_max_plot}},
  \code{\link{plot_3d_landscape}},
  \code{\link{plot_3d_profiles}},
  \code{\link{plot_active}},
  \code{\link{plot_features_profiles}},
  \code{\link{plot_features_profile}},
  \code{\link{plot_profiles}}, \code{\link{plot_profile}}
}
