% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_generation.R
\name{aggregate_profiles}
\alias{aggregate_profiles}
\title{Aggregates multiple heart rate profiles}
\usage{
aggregate_profiles(profiles)
}
\arguments{
\item{profiles}{List of profiles to be aggregated
See \code{\link{get_profile}} for a description of profile.}
}
\value{
List of means and standard deviations of heart rate at all data points.
}
\description{
Finds mean and standard deviation of multiple heart rate profiles
}
\examples{
profiles <- get_profiles(df, 500, 500, 500)
aggregate_profiles(profiles)

}
\seealso{
\code{\link{get_profiles}}

Other profile functions: \code{\link{get_profiles}},
  \code{\link{get_profile}}, \code{\link{normalise}}
}
