% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_extraction.R
\name{all_start_end_points}
\alias{all_start_end_points}
\title{Find start and end points of all activities}
\usage{
all_start_end_points(energy_df, time_threshold = 15, rest_threshold = 5)
}
\arguments{
\item{energy_df}{An energy dataframe as returned by \code{\link{load_tcx_file}}.}

\item{time_threshold}{The minumum number of seconds duration for an activity.}

\item{rest_threshold}{The maximum number of resting seconds within and activity.}
}
\value{
List containing :
\itemize{
\item data frame of all the starting and ending indeces of the activities.
\item energy data frame where missing data points have been interpolated.
}
}
\description{
find index for start and end of all active periods in an energy data frame
}
\seealso{
\code{\link{load_tcx_file}} for format of input dataframe

Other get activity functions: \code{\link{get_activities}},
  \code{\link{start_end_points}}
}
