% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{coefficients}
\alias{coefficients}
\title{Gets coefficients of fitted polynomial polynomial}
\usage{
coefficients(profile, start, end, degree, model_out = FALSE)
}
\arguments{
\item{profile}{profile to be nomalised
See \code{\link{get_profile}} for a description of profile.}

\item{start}{Index of start of region to fit polynomial}

\item{end}{Index of end of region to fit polynomial}

\item{degree}{Degree of polynomial}

\item{model_out}{Return entire model else just return the coefficients}
}
\value{
coefficients of the model fit the region of the profile
}
\description{
Gets coefficients of fitted polynomial polynomial
}
\seealso{
\code{\link{get_profile}} for a description of profile.

Other feature functions: \code{\link{get_features}},
  \code{\link{hr_change}}, \code{\link{max_hr}},
  \code{\link{resting_hr}}, \code{\link{resting_indeces}},
  \code{\link{slope}}
}
