% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_color}
\alias{get_color}
\title{Given an index returns a plotly color}
\usage{
get_color(i)
}
\arguments{
\item{i}{Index of color}
}
\value{
Color to be used in plotting function
}
\description{
returns a plotly color to be used in plotting functions
}
\seealso{
Other plotting functions: \code{\link{add_sections}},
  \code{\link{get_energy_profiles}},
  \code{\link{min_max_plot}},
  \code{\link{plot_3d_profiles}},
  \code{\link{plot_active}},
  \code{\link{plot_features_profiles}},
  \code{\link{plot_features_profile}},
  \code{\link{plot_profiles}}, \code{\link{plot_profile}}
}
