% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_energy_profiles}
\alias{get_energy_profiles}
\title{Get profiles from energy dataframes with added information ready for plotting
with plot_3d_profiles}
\usage{
get_energy_profiles(energy_dfs, buffer_start, buffer_end, scale,
  time_threshold = 15)
}
\arguments{
\item{energy_dfs}{list of profiles or aggregate profiles to be plotted
See \code{\link{get_profile}} for a description of profile.
See \code{\link{aggregate_profiles}} for a description of an aggregate profile.}

\item{buffer_start}{The number of seconds before activity included in profile}

\item{buffer_end}{The number of seconds after activity included in profile}

\item{scale}{The number of data points in each third of the final profile}
}
\value{
List of dataframes with profiles and other useful information
}
\description{
prepare data to be plotted by plot_3d_profiles
}
\seealso{
\code{\link{get_profile}} for a description of profile.
    \code{\link{aggregate_profiles}} for a description of an aggregate profile.

Other plotting functions: \code{\link{add_sections}},
  \code{\link{get_color}}, \code{\link{min_max_plot}},
  \code{\link{plot_3d_landscape}},
  \code{\link{plot_3d_profiles}},
  \code{\link{plot_active}},
  \code{\link{plot_features_profiles}},
  \code{\link{plot_features_profile}},
  \code{\link{plot_profiles}}, \code{\link{plot_profile}}
}
