% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_generation.R
\name{get_profile}
\alias{get_profile}
\title{Convert energy data frame into heart rate profile}
\usage{
get_profile(energy_df, buffer_start, buffer_end, scale, rest_sec = 180,
  points = NULL)
}
\arguments{
\item{energy_df}{A dataframe with "energy" format as returned by \code{\link{load_tcx_file}}}

\item{buffer_start}{The number of seconds before activity included in profile}

\item{buffer_end}{The number of seconds after activity included in profile}

\item{scale}{The number of data points in each third of the final profile}

\item{rest_sec}{The number of seconds of recovery included in initial recovery period}

\item{points}{A list with start and end points of activities. If NULL then it will
use \code{\link{start_end_points}} to caluculate the start and end points of
the activity.}
}
\value{
Named list of heart rates (bpm) scaled into profile consiting of 3 regions :
\itemize{
\item resting period before activity
\item activity and initial recovery period
\item recovery after activity and inital recovery
}
}
\description{
scales final profile into 3 regions. If unable to create profile from energy data
     frame nothing is returned and a warning supplied.
}
\examples{
get_profile(df, 500, 500, 500)
get_profile(df, 200, 200, 500, rest_sec = 180)
get_profile(df, 100, 200, 500,
    points = list(
        "start_point" = 600,
        "end_point" = 1000
    )
)

}
\seealso{
\code{\link{load_tcx_file}} for format of input dataframe,
    \code{\link{start_end_points}} for more on start and end points of activities.

Other profile functions: \code{\link{aggregate_profiles}},
  \code{\link{get_profiles}}, \code{\link{normalise}}
}
