% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_generation.R
\name{get_profiles}
\alias{get_profiles}
\title{Converts multiple energy dataframes into heart rate profiles}
\usage{
get_profiles(energy_dfs, buffer_start, buffer_end, scale, rest_sec = 180,
  norm = FALSE, norm_resting = FALSE, all_activities = TRUE,
  time_threshold = 15, rest_threshold = 5)
}
\arguments{
\item{energy_dfs}{list of energy data frames such as returned by \code{\link{load_tcx_file}}}

\item{buffer_start}{The number of seconds before activity included in profile}

\item{buffer_end}{The number of seconds after activity included in profile}

\item{scale}{The number of data points in each third of the final profile}

\item{rest_sec}{The number of seconds in initial resting part of final profile. Must be less
than buffer_end}

\item{norm}{Normalise profile heart rates between 0 and 1}

\item{all_activities}{Finds profiles for all activities in the energy dataframes or else
finds only the first activity in each dataframe}

\item{time_threshold}{The minumum number of seconds duration for an activity}

\item{rest_threshold}{The maximum number of resting seconds within and activity}
}
\value{
List containing profiles
     See \code{\link{get_profile}} for a description of profile.
}
\description{
Scales final profiles into 3 regions. If unable to create profiles from some
dataframes the profile is not included and a warning is supplied.
}
\examples{
get_profiles(df, 500, 500, 500)
get_profiles(df, 200, 200, 1000, norm = TRUE)
get_profiles(df, 1500, 500, 100, rest_sec = 80, all_activities = FALSE)

}
\seealso{
\code{\link{load_tcx_dir}} for a description of an energy dataframe.

Other profile functions: \code{\link{aggregate_profiles}},
  \code{\link{get_profile}}, \code{\link{normalise}}
}
