% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{hr_change}
\alias{hr_change}
\title{Get change between resting heart rate and final heart rate}
\usage{
hr_change(profile)
}
\arguments{
\item{profile}{Profile to find heart rate change.
See \code{\link{get_profile}} for a description of profile.}
}
\value{
Returns a value of the change between resting and final heart rate
}
\description{
gets the difference between the average resting and the final heart rate of a profile.
}
\seealso{
\code{\link{get_profile}} for a description of profile.

Other feature functions: \code{\link{get_features}},
  \code{\link{max_hr}}, \code{\link{profile_coefficients}},
  \code{\link{resting_hr}}, \code{\link{resting_indeces}},
  \code{\link{slope}}
}
