% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcx_loading.R
\name{load_tcx_dir}
\alias{load_tcx_dir}
\title{Load all tcx files in a directory and load into a list of energy dataframes}
\usage{
load_tcx_dir(directory, recursive = TRUE)
}
\arguments{
\item{directory}{A filepath to a directory}

\item{recursive}{Boolean to determine if data loading is recursively or not}
}
\value{
Named list of energy dataframes where the name is the path to the tcx file.
    See \code{\link{load_tcx_file}} for a description of an energy dataframe.
}
\description{
Load all tcx files in a directory and load into a list of energy dataframes
}
\examples{
load_tcx_dir("topDir/secondLevel/")
load_tcx_dir("../")
load_tcx_dir(".", recursive = FALSE)
}
\seealso{
\code{\link{load_tcx_file}} for loading multiple tcx files

Other load tcx functions: \code{\link{load_tcx_file}},
  \code{\link{merge_list}}, \code{\link{select_list}},
  \code{\link{selectdf}}
}
