% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcx_loading.R
\name{load_tcx_file}
\alias{load_tcx_file}
\title{Process a single tcx file and return and energies dataframe.}
\usage{
load_tcx_file(filename)
}
\arguments{
\item{filename}{A filepath to a tcx file (.tcx)}
}
\value{
The energies dataframe from the file at \code{filename}
energy dataframe :
\itemize{
\item totalE  : total energy expenditure
\item deltaE  : delta energy expenditure
\item bpm     : heart rate in beats per minute (bpm)
\item deltaBPM: dela heart rate in beats per minute (bpm)
\item time    : POSIXct timestamps
\item rawbpm  : raw unsmoothed heart rate}
}
\description{
Process a single tcx file and return and energies dataframe.
}
\examples{
load_tcx_file("topDir/secondLevel/fitbitdata.tcx")
load_tcx_file("./fitbitdata.tcx")
load_tcx_file("fitbitdata.tcx")
}
\references{
Energy expenditure calculations used are described in:

   Weyand, P. G., Smith, B. R., Puyau, M. R., & Butte, N. F. (2010).
   The mass-specific energy cost of human walking is set by stature.
   The Journal of Experimental Biology, 213(Pt 23), 3972–3979.

   Kawata, A., Shiozawa, N., & Makikawa, M. (2007).
   Estimation of Energy Expenditure during Walking Including UP/Down Hill.
   In R. Magjarevic & J. H. Nagel (Eds.), World Congress on Medical Physics And Biomedical Engineering 2006,
   441-444.
}
\seealso{
\code{\link{load_tcx_dir}} for loading multiple tcx files

Other load tcx functions: \code{\link{load_tcx_dir}},
  \code{\link{merge_list}}, \code{\link{select_list}},
  \code{\link{selectdf}}
}
