% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{max_hr}
\alias{max_hr}
\title{Find maximum heart rate}
\usage{
max_hr(profile)
}
\arguments{
\item{profile}{profile used to find maximum heart rate
See \code{\link{get_profile}} for a description of profile.}
}
\value{
value of the maximum heart rate in the profile
}
\description{
find the maximum heart rate in the profile
}
\seealso{
\code{\link{get_profile}} for a description of profile.

Other feature functions: \code{\link{coefficients}},
  \code{\link{get_features}}, \code{\link{hr_change}},
  \code{\link{resting_hr}}, \code{\link{resting_indeces}},
  \code{\link{slope}}
}
