% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcx_loading.R
\name{merge_list}
\alias{merge_list}
\title{Merges a list of energy dataframes into one single dataframe}
\usage{
merge_list(dflist)
}
\arguments{
\item{dflist}{A list of dataframes all with the same column names and
with a time column for sorting}
}
\value{
a single dataframe
}
\description{
merges a list of energy dataframes into one single dataframe
}
\seealso{
Other load tcx functions: \code{\link{load_tcx_dir}},
  \code{\link{load_tcx_file}}, \code{\link{select_list}},
  \code{\link{selectdf}}
}
