% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_generation.R
\name{normalise}
\alias{normalise}
\title{Normalise heart rate in profile}
\usage{
normalise(profile)
}
\arguments{
\item{profile}{profile to be nomalised
See \code{\link{get_profile}} for a description of profile.}
}
\value{
profile df that has been normalised
    See \code{\link{get_profile}} for a description of profile.
}
\description{
normalises heart rate between 0 and 1 in heart rate dynamic profile
}
\seealso{
See \code{\link{get_profile}} for a description of profile.

Other profile functions: \code{\link{aggregate_profiles}},
  \code{\link{get_profiles}}, \code{\link{get_profile}}
}
