% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_3d_landscape}
\alias{plot_3d_landscape}
\title{Plot multiple profiles in 3d landscape based on energy}
\usage{
plot_3d_landscape(energy_dfs, buffer_start, buffer_end, scale,
  time_threshold = 15, plot_energy = FALSE)
}
\arguments{
\item{energy_dfs}{list of profiles or aggregate profiles to be plotted
See \code{\link{get_profile}} for a description of profile.
See \code{\link{aggregate_profiles}} for a description of an aggregate profile.}

\item{buffer_start}{The number of seconds before activity included in profile}

\item{buffer_end}{The number of seconds after activity included in profile}

\item{scale}{The number of data points in each third of the final profile}
}
\value{
Plotly plot
}
\description{
plot multiple profiles in 3d landscape by interpolating heart rate data points
}
\examples{
plot_3d_landscape(energy_dfs, buffer_start, buffer_end, scale)
}
\seealso{
\code{\link{load_tcx_file}} for an explanation of energy dataframes.

Other plotting functions: \code{\link{add_sections}},
  \code{\link{get_color}},
  \code{\link{get_energy_profiles}},
  \code{\link{min_max_plot}},
  \code{\link{plot_3d_profiles}},
  \code{\link{plot_active}},
  \code{\link{plot_features_profiles}},
  \code{\link{plot_features_profile}},
  \code{\link{plot_profiles}}, \code{\link{plot_profile}}
}
