% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_active}
\alias{plot_active}
\title{Plot all the activities}
\usage{
plot_active(energy_df, p = plot_ly(), time_threshold = 15,
  rest_threshold = 5, type = "bpm", real_time = TRUE, plot_all = TRUE)
}
\arguments{
\item{energy_df}{A dataframe with "energy" format as returned by \code{\link{load_tcx_file}}.}

\item{p}{A plot to be added to.}

\item{time_threshold}{The minumum number of seconds duration for an activity.}

\item{rest_threshold}{The maximum number of resting seconds within and activity.}

\item{type}{The type of data plotted:
\itemize{
\item "bpm" : plots the heart rate data in bpm
\item "deltaE" : plots the change in energy expenditure
\item "totalE" : plots the total energy expenditure
}}

\item{real_time}{Plots data in real time or else uses indexes.}

\item{plot_all}{Plots the activities and all the data from the data frame.}
}
\value{
A plotly object.
}
\description{
finds and plots all the activities in a energy data frame
}
\seealso{
\code{\link{load_tcx_file}} for details on input format.

Other plotting functions: \code{\link{add_sections}},
  \code{\link{get_color}},
  \code{\link{get_energy_profiles}},
  \code{\link{min_max_plot}},
  \code{\link{plot_3d_landscape}},
  \code{\link{plot_3d_profiles}},
  \code{\link{plot_features_profiles}},
  \code{\link{plot_features_profile}},
  \code{\link{plot_profiles}}, \code{\link{plot_profile}}
}
