% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_features_profiles}
\alias{plot_features_profiles}
\title{Plot features of multiple profiles}
\usage{
plot_features_profiles(profiles, labels, buffer_end)
}
\arguments{
\item{profiles}{list of profiles or aggregate profiles to be plotted
See \code{\link{get_profile}} for a description of profile.
See \code{\link{aggregate_profiles}} for a description of an aggregate profile.}

\item{buffer_end}{The number seconds before activity included in profile.}
}
\value{
Plotly plot
}
\description{
plot the features of all the profiles in a list on the same plot,
   to easily compare between features from different profiles
}
\examples{
   profiles <- get_profiles(dfs, 200, 300, 500)
   plot_features_profiles(profiles, names(profiles), 300)

   agg_profile1 <- aggregate_profiles(select(profiles, "1"))
   agg_profile2 <- aggregate_profiles(select(profiles, "2"))
   plot_features_profiles(list(agg_profile1, agg_profile2), list("1", "2"), 300)
}
\seealso{
\code{\link{get_profile}} for a description of profile.
    \code{\link{aggregate_profiles}} for a description of an aggregate profile.

Other plotting functions: \code{\link{add_sections}},
  \code{\link{get_color}},
  \code{\link{get_energy_profiles}},
  \code{\link{min_max_plot}},
  \code{\link{plot_3d_profiles}},
  \code{\link{plot_active}},
  \code{\link{plot_features_profile}},
  \code{\link{plot_profiles}}, \code{\link{plot_profile}}
}
