% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_profile}
\alias{plot_profile}
\title{Plot a single profile}
\usage{
plot_profile(profile, label, title = NULL, color = get_color(0),
  p = plot_ly(), sections = "shade", sec_opacity = 0.3,
  ribbon_opactiy = 0.4)
}
\arguments{
\item{profile}{Profile or aggregate profile to be plotted.
See \code{\link{get_profile}} for a description of profile.
See \code{\link{aggregate_profiles}} for a description of an aggregate profile.}

\item{label}{Label of the profile being plotted}

\item{title}{Title of the plot}

\item{color}{Color of the profile added to the plot.}

\item{p}{Plot to be added to, if none provided a new plot will be created.}

\item{sections}{Style used to plot sections of the profile.
   See \code{\link{get_profile}} for description of profile sections.

   The style options are:
   \itemize{
   \item "shade" : shade the background of the different sections different colors
   \item "lines" : add lines to separate the different sections
   }}

\item{sec_opacity}{The opacity of the sections of the profile added to the plot.}

\item{ribbon_opacity}{The opacity of standard deviation ribbon when plotting aggregate profile.
See \code{\link{aggregate_profiles}} for a description of an aggregate profile.}
}
\value{
Plotly object.
}
\description{
plot single profiles individually or add to existing plot
}
\examples{
   profile1 <- get_profile(df, 500, 500, 500)
   p <- plot_profile(profile1, "profile 1", title = "Great Profiles")
   profile2 <- get_profile(df, 200, 300, 500)
   plot_profile(profile2, "profile 2", p = p)

   profiles <- get_profiles(dfs, 500, 500, 500)
   agg_profile <- aggregate_profiles(profiles)
   plot_profile(agg_profile, "aggregated profile", ribbon_opacity = 0.2)
}
\seealso{
\code{\link{get_profile}} for a description of profile.
    \code{\link{aggregate_profiles}} for a description of an aggregate profile.

Other plotting functions: \code{\link{add_sections}},
  \code{\link{get_color}},
  \code{\link{get_energy_profiles}},
  \code{\link{min_max_plot}},
  \code{\link{plot_3d_profiles}},
  \code{\link{plot_active}},
  \code{\link{plot_features_profiles}},
  \code{\link{plot_features_profile}},
  \code{\link{plot_profiles}}
}
