% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_profiles}
\alias{plot_profiles}
\title{Plot a many individual profiles}
\usage{
plot_profiles(profiles, labels, title = NULL, p = plot_ly(),
  sections = "shade", sec_opacity = 0.3)
}
\arguments{
\item{profiles}{A list of profiles or aggregate profiles to be plotted.
See \code{\link{get_profile}} for a description of profile.
See \code{\link{aggregate_profiles}} for a description of an aggregate profile.}

\item{labels}{A list of labels of the profiles being plotted.}

\item{title}{Title of the plot}

\item{p}{Plot to be added to, if none provided a new plot will be created.}

\item{sections}{Style used to show sections of the profile.
   See \code{\link{get_profile}} for description of profile sections.

   The style options are:
   \itemize{
   \item "shade" : shade the background of the different sections different colors
   \item "lines" : add lines to separate the different sections
   }}

\item{sec_opacity}{The opacity of the sections of the profile added to the plot.}
}
\value{
Plotly object.
}
\description{
plot many profiles or aggregate profiles on one plot
}
\examples{
   # Plot two profiles
   profile1 <- get_profile(df, 500, 500, 500)
   profile2 <- get_profile(df, 200, 300, 500)
   plot_profiles(list(profile1, profile2), list("profile 1" "profile 2"))

   # Plot many profiles
   profiles <- get_profiles(df, 200, 300, 500)
   plot_profiles(profiles, names(profiles), sections = "shade")

   # Plot aggregate profiles
   profiles1 <- get_profiles(dfs, 500, 500, 500)
   profiles2 <- get_profiles(df, 200, 300, 500)
   agg_profile1 <- aggregate_profiles(profiles1)
   agg_profile2 <- aggregate_profiles(profiles2)
   plot_profiles(list(agg_profile1, agg_profile2), list("aggregated profile 1",
                "aggregated profile 2"), title = "Aggregate profiles")

}
\seealso{
\code{\link{get_profile}} for a description of a profile.
    \code{\link{aggregate_profiles}} for a description of an aggregate profile.

Other plotting functions: \code{\link{add_sections}},
  \code{\link{get_color}},
  \code{\link{get_energy_profiles}},
  \code{\link{min_max_plot}},
  \code{\link{plot_3d_profiles}},
  \code{\link{plot_active}},
  \code{\link{plot_features_profiles}},
  \code{\link{plot_features_profile}},
  \code{\link{plot_profile}}
}
