% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{resting_hr}
\alias{resting_hr}
\title{Find average resting heart rate}
\usage{
resting_hr(profile)
}
\arguments{
\item{profile}{profile used to find resting heart rate
See \code{\link{get_profile}} for a description of profile.}

\item{scale}{The number of data points in each third of the final profile, used to find start of activity.}
}
\value{
value of the average resting heart rate
}
\description{
find the resting heart rate by averaging the resting period before activity
}
\seealso{
\code{\link{get_profile}} for a description of profile.

Other feature functions: \code{\link{coefficients}},
  \code{\link{get_features}}, \code{\link{hr_change}},
  \code{\link{max_hr}}, \code{\link{resting_indeces}},
  \code{\link{slope}}
}
