% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcx_loading.R
\name{select_list}
\alias{select_list}
\title{Select elements of a named list using a regex on the name}
\usage{
select_list(list, regex)
}
\arguments{
\item{list}{A named list of anything}

\item{regex}{A regular expression used to select elements}
}
\value{
Subset of the original list where the name matches \code{regex}
}
\description{
Select elements of a named list using a regex on the name
}
\examples{
select(list("a" = 1, "abba" = 2, "b" = 3), "a")
    $ a   : num 1
    $ abba: num 2
select(list("a" = 1, "abba" = 2, "b" = 3), "b")
    $ abba: num 2
    $ b   : num 3
select(list("a" = 1, "abba" = 2, "b" = 3), "[[:alpha:]]{4}")
    $ abba: num 2
}
\seealso{
Other load tcx functions: \code{\link{load_tcx_dir}},
  \code{\link{load_tcx_file}}, \code{\link{merge_list}},
  \code{\link{selectdf}}
}
