% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcx_loading.R
\name{selectdf}
\alias{selectdf}
\title{Select rows of a dataframe}
\usage{
selectdf(list, regex)
}
\arguments{
\item{regex}{A regular expression used to select elements}

\item{df}{A dataframe with named rows}
}
\value{
subset of the original dataframe where the rowname matches \code{regex}
}
\description{
select rows of a dataframe using a regex on the rownames
}
\examples{
df <- t(data.frame("a" = c(1:3), "abba" = c(2:4), "b" = c(3:5)))
df
         [,1] [,2] [,3]
    a       1    2    3
    abba    2    3    4
    b       3    4    5
selectdf(df, "a")
         [,1] [,2] [,3]
    a       1    2    3
    abba    2    3    4
selectdf(df, "b")
         [,1] [,2] [,3]
    abba    2    3    4
    b       3    4    5
selectdf(df, "[[:alpha:]]{4}")
    [1] 2 3 4
}
\seealso{
Other load tcx functions: \code{\link{load_tcx_dir}},
  \code{\link{load_tcx_file}}, \code{\link{merge_list}},
  \code{\link{select_list}}
}
