% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{slope}
\alias{slope}
\title{Fit line to recovery slope}
\usage{
slope(profile, model_out = FALSE)
}
\arguments{
\item{profile}{profile to be nomalised
See \code{\link{get_profile}} for a description of profile.}

\item{model_out}{Return entire model else just return the coefficients}

\item{scale}{The number of data points in each third of the final profile.
Used to find the start of activity}
}
\value{
coefficients of the model fit the region of the profile
}
\description{
fit regression line to the slope of the final recovery in a profile.
}
\seealso{
\code{\link{get_profile}} for a description of profile.

Other feature functions: \code{\link{coefficients}},
  \code{\link{get_features}}, \code{\link{hr_change}},
  \code{\link{max_hr}}, \code{\link{resting_hr}},
  \code{\link{resting_indeces}}
}
