% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_extraction.R
\name{start_end_points}
\alias{start_end_points}
\title{Finds the activity in an energy dataframe}
\usage{
start_end_points(energy_df, time_threshold = 15, rest_threshold = 5)
}
\arguments{
\item{energy_df}{An energy data frame as returned by \code{\link{load_tcx_file}}.}

\item{time_threshold}{The minimum number of seconds duration of an activity.}

\item{rest_threshold}{The maximum number of resting seconds within and activity.}
}
\value{
named list with \code{start_point} and \code{end_point} for the activity.
}
\description{
finds the first start and end points of activity in energy dataframe.
}
\examples{
start_end_points(df)
start_end_points(df, time_threshold = 30, rest_threshold = 2)
start_end_points(df, rest_threshold = 2)
}
\seealso{
\code{\link{load_tcx_file}} for format of input dataframe

Other get activity functions: \code{\link{all_start_end_points}},
  \code{\link{get_activities}}
}
