% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{get_features}
\alias{get_features}
\title{Find all the features in a profile}
\usage{
get_features(profiles, buffer_end, rest_sec = 180, degree_coef_active = 5,
  degree_coef_init_rec = 5)
}
\arguments{
\item{profiles}{list of profiles such as returned by \code{\link{get_profile}} used to find features}

\item{buffer_end}{The number seconds before activity included in profile.}

\item{rest_sec}{The number of seconds of recovery included in initial recovery period.}

\item{degree_coef_init_rec}{The degree of the polynomial fitted to initial recovery period of a profile
see \code{\link{coefficients}} for details on fitting the polynomial
and \code{\link{get_profile}} for a description of a profile}

\item{scale}{The number of data points in each third of the final profile.}

\item{degree_coeff_active}{The degree of the polynomial fitted to active period of a profile
see \code{\link{coefficients}} for details on fitting the polynomial
and \code{\link{get_profile}} for a description of a profile}
}
\value{
Data frame containing features and labels with features from each profile on a new row
}
\description{
find all the features of a profile
}
\examples{
profiles <- get_profiles(dfs, 200, 300, 500)
get_features(profiles, 300, 500)
}
\seealso{
\code{\link{get_profile}} for a description of a profile.
\code{\link{coefficients}} for details of fitting the polynomials.

Other feature functions: \code{\link{coefficients}},
  \code{\link{hr_change}}, \code{\link{max_hr}},
  \code{\link{resting_hr}}, \code{\link{resting_indeces}},
  \code{\link{slope}}
}
