% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{hr_change}
\alias{hr_change}
\title{Gets change between resting heart rate and final heart rate}
\usage{
hr_change(profile)
}
\arguments{
\item{profile}{profile to be nomalised
See \code{\link{get_profile}} for a description of profile.}

\item{scale}{The number of data points in each third of the final profile.
Used to find the start of activity}

\item{model_out}{Return entire model else just return the coefficients}
}
\value{
Returns a value of the change between resting and final heart rate
}
\description{
gets the difference between average resting and final heart rate for a profile.
}
\seealso{
\code{\link{get_profile}} for a description of profile.

Other feature functions: \code{\link{coefficients}},
  \code{\link{get_features}}, \code{\link{max_hr}},
  \code{\link{resting_hr}}, \code{\link{resting_indeces}},
  \code{\link{slope}}
}
