% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcx_loading.R
\name{load_tcx_dir}
\alias{load_tcx_dir}
\title{Process all tcx files in a directory}
\usage{
load_tcx_dir(directory, recursive = TRUE)
}
\arguments{
\item{directory}{A filepath to a directory.}

\item{recursive}{Process tcx files in all levels of the directory or only the top level.}
}
\value{
named list of energy dataframes where the name is the path to the tcx file.
    See \code{\link{load_tcx_file}} for a description of an energy dataframe.
}
\description{
process all tcx files in a directory and produce a list of energy dataframes.
}
\examples{
load_tcx_dir("topDir/secondLevel/")
load_tcx_dir("../")
load_tcx_dir(".", recursive = FALSE)
}
\seealso{
\code{\link{load_tcx_file}} for loading multiple tcx files

Other load tcx functions: \code{\link{load_tcx_file}},
  \code{\link{merge_list}}, \code{\link{select_list}},
  \code{\link{selectdf}}
}
