% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{slope}
\alias{slope}
\title{Fit line to recovery slope}
\usage{
slope(profile, model_out = FALSE)
}
\arguments{
\item{profile}{The profile used to find the recovery slope.
See \code{\link{get_profile}} for a description of profile.}

\item{model_out}{Return entire model or else just return the coefficients}
}
\value{
coefficients of the model fit the final recovery region of the profile
}
\description{
fit regression line to the slope of the final recovery in a profile.
}
\seealso{
\code{\link{get_profile}} for a description of profile.

Other feature functions: \code{\link{get_features}},
  \code{\link{hr_change}}, \code{\link{max_hr}},
  \code{\link{profile_coefficients}},
  \code{\link{resting_hr}}, \code{\link{resting_indeces}}
}
